engine: http_server: Make the flush_now_count atomic
The `flush_now_count` variable is accessed across threads (engine + HTTP server) and needs proper safety.

+++ include/fluent-bit/flb_config.h
@@ -21,6 +21,7 @@
 #define FLB_CONFIG_H
 
 #include <time.h>
+#include <stdint.h>
 
 #include <fluent-bit/flb_info.h>
 #include <fluent-bit/flb_pipe.h>
@@ -73,7 +74,7 @@ struct flb_config {
     int grace_count;              /* Count of grace shutdown tries           */
     int grace_input;              /* Shutdown grace to keep inputs ingesting */
     flb_pipefd_t flush_fd;        /* Timer FD associated to flush            */
-    unsigned int flush_now_count; /* count of on-demand flush requests       */
+    uint64_t flush_now_count;     /* count of on-demand flush requests       */
     int convert_nan_to_null;      /* Convert null to nan ?                   */
 
     int daemon;               /* Run as a daemon ?              */
